Finished Blog module and correct a number of bugs
Completed blog module
Added Sql functions for datetime to Helper
Fixed small isues across platform
Minor fix to material template

+++ app/modules/arcblog/controller/deletecategory.php
@@ -2,6 +2,14 @@
 
 if (system\Helper::arcIsAjaxRequest()) {
         $category = BlogCategory::getByID($_POST["id"]);
+
+        $blogs = Blog::getAllByCategoryID($_POST["id"]);
+        if (count($blogs) > 0) {
+                system\Helper::arcAddMessage("danger", "Unable to delete category with attached posts");
+                system\Helper::arcReturnJSON(["error" => "Unable to delete category with attached posts"]);
+                return;
+        }
+
         $category->delete($_POST["id"]);
 
         $currentRoute = Router::getRoute("blog/category/" . $category->seourl);
